Openstack backoff logic
